# generated/sdk-java/api/openapi.yaml
openapi: 3.0.3
info:
  description: "RESTful API for inventory and product management, secured with JWT."
  title: Inventory Management API
  version: 1.0.0
servers:
- url: http://localhost:8080
security:
- bearerAuth: []
tags:
- name: Authentication
- name: Inventories
- name: Products
paths:
  /auth/register:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Invalid request
        "409":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Conflict
      security: []
      summary: User registration
      tags:
      - Authentication
      x-content-type: application/json
      x-accepts: application/problem+json
  /auth/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Invalid request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Unauthorized
      security: []
      summary: User login
      tags:
      - Authentication
      x-content-type: application/json
      x-accepts: application/problem+json
  /api/inv:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Inventory'
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Invalid request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Unauthorized
      summary: Create inventory
      tags:
      - Inventories
      x-content-type: application/json
      x-accepts: application/problem+json
  /api/inv/paginated:
    get:
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 0
          type: integer
        style: form
      - explode: true
        in: query
        name: size
        required: false
        schema:
          default: 10
          type: integer
        style: form
      responses:
        "200":
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Invalid request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Unauthorized
      summary: List inventories (Paginated)
      tags:
      - Inventories
      x-accepts: application/problem+json
  /api/inv/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Invalid request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Unauthorized
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Resource not found
      summary: Delete inventory
      tags:
      - Inventories
      x-accepts: application/problem+json
  /api/productos:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProductDto'
                type: array
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Invalid request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Unauthorized
      summary: List all products
      tags:
      - Products
      x-accepts: "application/json,application/problem+json"
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Invalid request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Unauthorized
      summary: Add standalone product
      tags:
      - Products
      x-content-type: application/json
      x-accepts: application/problem+json
  /api/productos/paginated:
    get:
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 0
          type: integer
        style: form
      - explode: true
        in: query
        name: size
        required: false
        schema:
          default: 10
          type: integer
        style: form
      responses:
        "200":
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Invalid request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Unauthorized
      summary: List all products (Paginated)
      tags:
      - Products
      x-accepts: application/problem+json
  /api/productos/{product_id}:
    get:
      parameters:
      - explode: false
        in: path
        name: product_id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Invalid request
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Resource not found
      summary: Get product by ID
      tags:
      - Products
      x-accepts: application/problem+json
  /api/productos/inv/{inventory_id}:
    post:
      parameters:
      - explode: false
        in: path
        name: inventory_id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Invalid request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Unauthorized
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Resource not found
      summary: Add product to specific inventory
      tags:
      - Products
      x-content-type: application/json
      x-accepts: application/problem+json
  /api/productos/inv/{inventory_id}/paginated:
    get:
      parameters:
      - explode: false
        in: path
        name: inventory_id
        required: true
        schema:
          type: integer
        style: simple
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 0
          type: integer
        style: form
      - explode: true
        in: query
        name: size
        required: false
        schema:
          default: 10
          type: integer
        style: form
      responses:
        "200":
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Invalid request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Unauthorized
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Resource not found
      summary: List products from inventory (Paginated)
      tags:
      - Products
      x-accepts: application/problem+json
  /api/productos/inv/{inventory_id}/{product_id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: inventory_id
        required: true
        schema:
          type: integer
        style: simple
      - explode: false
        in: path
        name: product_id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          description: OK
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Unauthorized
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Resource not found
      summary: Delete product from inventory
      tags:
      - Products
      x-accepts: application/problem+json
    put:
      parameters:
      - explode: false
        in: path
        name: inventory_id
        required: true
        schema:
          type: integer
        style: simple
      - explode: false
        in: path
        name: product_id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Invalid request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Unauthorized
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Resource not found
      summary: Update product in inventory
      tags:
      - Products
      x-content-type: application/json
      x-accepts: application/problem+json
components:
  responses:
    BadRequest:
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
      description: Invalid request
    Unauthorized:
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
      description: Unauthorized
    NotFound:
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
      description: Resource not found
    Conflict:
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
      description: Conflict
  schemas:
    ProblemDetail:
      properties:
        type:
          example: about:blank
          type: string
        title:
          example: Resource not found
          type: string
        status:
          example: 404
          type: integer
        detail:
          example: Product with id 10 not found
          type: string
        instance:
          example: /api/productos/10
          type: string
      type: object
    RegisterRequest:
      example:
        password: password
        role: role
        username: username
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
      type: object
    LoginRequest:
      example:
        password: password
        username: username
      properties:
        username:
          type: string
        password:
          type: string
      type: object
    Product:
      example:
        price: 6.027456183070403
        name: name
        weight: 1.4658129805029452
        id: 0
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          format: double
          type: number
        weight:
          format: double
          type: number
      required:
      - name
      - price
      type: object
    ProductDto:
      example:
        price: 0.8008281904610115
        name: name
        weight: 6.027456183070403
      properties:
        name:
          type: string
        price:
          format: double
          type: number
        weight:
          format: double
          type: number
      type: object
    Inventory:
      example:
        name: name
        id: 0
      properties:
        id:
          type: integer
        name:
          type: string
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http